(SIMP-2733) Check for gdm_version
gdm_version is set to undef on the first run, while the gdm package is
installed. After the first run, the file that the gdm_version fact is
confined to exists. Now the logic that includes versioncmp will check
for the existence of gdm_version before running the versioncmp.

SIMP-2733 #close